// apps/web/src/shared/api/generated/model/lessonResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Gravit API Docs
 * 앱센터 16.5기 동계 프로젝트 Gravit API Docs
 * OpenAPI spec version: 1.0.0
 */
import type { ProblemResponse } from './problemResponse';
import type { UnitSummaryResponse } from './unitSummaryResponse';

/**
 * 레슨 조회 Response
 */
export interface LessonResponse {
  /** 유닛 요약 정보 */
  unitSummaryResponse: UnitSummaryResponse;
  /** 문제 목록 */
  problems: ProblemResponse[];
  /** 전체 문제 수 */
  totalProblems: number;
}